Move CommandCenterLog to sentinel-transport-common and polish related code (#1341)
