+++ sentinel-transport/sentinel-transport-simple-http/src/main/java/com/alibaba/csp/sentinel/transport/heartbeat/client/SimpleHttpClient.java
@@ -25,7 +25,6 @@ import java.nio.charset.Charset;
 import java.util.Map;
 import java.util.Map.Entry;
 
-import com.alibaba.csp.sentinel.log.CommandCenterLog;
 import com.alibaba.csp.sentinel.log.RecordLog;
 
 /**
@@ -125,7 +124,7 @@ public class SimpleHttpClient {
                 try {
                     socket.close();
                 } catch (Exception ex) {
-                    CommandCenterLog.info("Error when closing " + type + " request to " + socketAddress + ": ", ex);
+                    RecordLog.warn("Error when closing {} request to {} in SimpleHttpClient", type, socketAddress, ex);
                 }
             }
         }
@@ -177,7 +176,7 @@ public class SimpleHttpClient {
             }
             return paramsBuilder.toString();
         } catch (Throwable e) {
-            RecordLog.info("Encode request params fail", e);
+            RecordLog.warn("Encode request params fail", e);
             return "";
         }
     }
